src/main/kotlin/com/wutsi/application/web/service/FeignAcceptLanguageInterceptor.kt
package com.wutsi.application.web.service

import feign.RequestInterceptor
import feign.RequestTemplate
import org.springframework.http.HttpHeaders
import org.springframework.stereotype.Service
import javax.servlet.http.HttpServletRequest

@Service
class FeignAcceptLanguageInterceptor(
    private val request: HttpServletRequest
) : RequestInterceptor {
    override fun apply(template: RequestTemplate) {
        val language = request.getHeader(HttpHeaders.ACCEPT_LANGUAGE)
        if (language != null)
            template.header(HttpHeaders.ACCEPT_LANGUAGE, language)
    }
}
